backend/requirements_api.txt
# Kritikos API - Dependências

# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic para validação
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS (FastAPI já tem embutido)

# Cache e performance
redis==5.0.1
hiredis==2.2.3

# Logs e monitoramento
structlog==23.2.0
prometheus-client==0.19.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Documentação
mkdocs==1.5.3
mkdocs-material==9.4.8

# Utilitários
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# HTTP client
aiohttp==3.9.1

# Background tasks
celery==5.3.4
flower==2.0.1

# Rate limiting
slowapi==0.1.9

# Validation
email-validator==2.1.0
phonenumbers==8.13.25

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
